+++ lib/src/models/index.dart
@@ -1,3 +1,4 @@
+export 'cast_translations.dart';
 export 'chewie_chapter.dart';
 export 'option_item.dart';
 export 'options_translation.dart';